Fix consistency properties (#284)
* Fix consistency properties

"x-ms-blob-content-md5" header is for "PUT Blob" when "content-md5" is for
"GET Blob Properties", and is base64 encoded.

* XML deserialization of new consistency structures

* Add integration tests for consistency data

* use bytes

* back to StringDeserializer

* allow getting a fixed width slice

+++ sdk/storage/src/core/errors.rs
@@ -63,9 +63,9 @@ pub enum Error {
     #[error("UnexpectedXMLError: {0}")]
     UnexpectedXMLError(String),
     #[error("digest length {0} bytes instead of 16")]
-    DigestNot16BytesLong(u64),
+    DigestNot16BytesLong(usize),
     #[error("CRC64 length {0} bytes instead of 8")]
-    CRC64Not8BytesLong(u64),
+    CRC64Not8BytesLong(usize),
     #[error("At least one of these headers must be present: {0:?}")]
     HeadersNotFound(Vec<String>),
 }
